[指标体系服务]添加业务指标相关接口

+++ qk-dam/qike-dm-indicator/src/main/java/com/qk/dm/indicator/repositories/IdcBusinessRepository.java
@@ -3,6 +3,6 @@ package com.qk.dm.indicator.repositories;
 import com.qk.dam.jpa.base.BaseRepository;
 import com.qk.dm.indicator.entity.IdcBusiness;
 
-public interface IdcBusinessRepository extends BaseRepository<IdcBusiness, Integer> {
+public interface IdcBusinessRepository extends BaseRepository<IdcBusiness, Long> {
 
 }
